TestModels/dafny-dependencies/StandardLibrary/runtimes/python/src/standard_library/extern/pyproject.toml
[tool.poetry]
name = "standard-library-externs"
version = "0.1.0"
description = ""
authors = ["AWS <aws-arg-dafny@amazon.com>"]
packages = [
    # Globally install all internaldafny modules
    # such that a module X can be imported with `import X`
    # from anywhere in the Python runtime that installs this module
    { include = "*.py" },
]
# Include all of the following .gitignored files in package distributions,
# even though it is not included in version control
include = ["*.py"]

[tool.poetry.dependencies]
python = "^3.11.0"
standard-library-internaldafny = { path = "../internal_generated_dafny", develop = false}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"